Task 2

+++ python-hello_world/4-print_float.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python3
 number = 3.14159
-if isinstance(number, int):
+if isinstance(number, float):
     print(f"Float: {number:.2f}")
 else:
     print(f"Must be number")
